live share 08/09

+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/AumentoRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.Aumento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AumentoRepository extends JpaRepository<Aumento, Long> {
+public interface AumentoRepository extends JpaRepository<Aumento, Integer> {
 }
