+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/DocumentoRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.Documento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentoRepository extends JpaRepository<Documento, Long> {
+public interface DocumentoRepository extends JpaRepository<Documento, Integer> {
 }
