+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/PrecioRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.Precio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrecioRepository extends JpaRepository<Precio, Long> {
+public interface PrecioRepository extends JpaRepository<Precio, Integer> {
 }
