+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/VisitasRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.Visitas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VisitasRepository extends JpaRepository<Visitas, Long> {
+public interface VisitasRepository extends JpaRepository<Visitas, Integer> {
 }
